wip: improve session and guest handling logic

+++ features/profile/data/src/commonMain/kotlin/com/msoula/hobbymatchmaker/features/profile/data/dataSources/mappers/RemoteMappers.kt
@@ -11,7 +11,7 @@ fun UserProfileDomainModel.toUserProfileRemoteDataModel(): UserProfileRemoteData
 
 fun UserProfileRemoteDataModel.toUserProfileDomainModel(): UserProfileDomainModel {
     return UserProfileDomainModel(
-        uid = this.uid.toLong(),
+        uid = this.uid,
         name = "",
         avatarUrl = "",
         bio = "",
